pyproject.toml
[project]
name = "seju-lite"
version = "0.1.0"
description = "A lightweight Telegram-first AI assistant inspired by nanobot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=22.0,<23.0",
    "pydantic>=2.8,<3.0",
    "httpx>=0.27,<1.0",
    "orjson>=3.10,<4.0",
    "typer>=0.12,<1.0",
    "loguru>=0.7,<1.0",
    "python-dotenv>=1.0,<2.0",
    "google-genai>=1.68.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.6,<1.0",
    "mypy>=1.10,<2.0"
]

[project.scripts]
seju-lite = "seju_lite.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.hatch.build.targets.wheel]
packages = ["src/seju_lite"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.15.6",
]
